Client for Dataverse Repositories

Provides access to Dataverse 4+ APIs, enabling data search, retrieval, and deposit.

Details

Dataverse is open-source data repository management software developed by the Institute for Quantitative Social Science at Harvard University. This package provides an R interface to Dataverse version 4 repositories, including the principal Dataverse hosted at Harvard (https://dataverse.harvard.edu/). Users can use the package to search for data stored in a Dataverse repository, retrieve data and other files, and also use the package to directly create and archive their own research data and software.

A Dataverse is structured as a nested set of “dataverse” repositories, such that a single dataverse can contain “datasets” (a set of code files, data files, etc.) or other dataverses. Thus, users may want to search for dataverses (sets of dataverses and datasets), datasets (sets of files), or individual files, and retrieve those objects accordingly. To retrieve a given file, a user typically needs to know what dataset it is stored in. All datasets are identified by a persistent identifier (such as an DOI or Handle, depending on the age of the dataset and what Dataverse repository it is hosted in).

This package provides five main sets of functions to interact with Dataverse:

Search: dataverse_search

Data download: get_dataframe_by_name, get_dataverse, dataverse_contents, get_dataset, dataset_metadata, get_file

Data archiving (SWORD API): service_document, list_datasets, initiate_sword_dataset, delete_sword_dataset, publish_sword_dataset, add_file, delete_file

Dataverse management “native” API: create_dataverse, publish_dataverse, delete_dataverse

Dataset management “native” API: create_dataset, update_dataset, publish_dataset, delete_dataset, dataset_files, dataset_versions

References

Documentation for this R Package

Code Repository for the R Package

Dataverse API Documentation

Dataverse Homepage

Harvard IQSS Dataverse

See also

Author

Maintainer: Shiro Kuriwaki shirokuriwaki@gmail.com (ORCID)

Authors:

Will Beasley wibeasley@hotmail.com (ORCID)

Thomas J. Leeper thosjleeper@gmail.com (ORCID)

Philip Durbin philipdurbin@gmail.com (ORCID)

Sebastian Karcher karcher@u.northwestern.edu (ORCID)

Other contributors:

Jan Kanis [contributor]

Edward Jee [contributor]

Johannes Gruber JohannesB.Gruber@gmail.com (ORCID) [contributor]

Martin Morgan mtmorgan.xyz@gmail.com (ORCID) [contributor]